Initial Commit
This is the initial version of a Mac app for Gambatte, which uses native Cocoa APIs, and doesn’t need to rely on a third-party framework like QT.

This version is relatively basic; sound and video emulation works just fine, but more advanced features and customisation options aren’t available yet.

Includes a make script for libgambatte, which should really be moved pronto.