Enable console to be selected
This enables the menu options for selecting if you’d like to emulate as an original GB, a GBC, or a GBA.

At present, however, libgambatte does not support modification of the required LoadFlags without reloading the ROM file. Since this could cause problems (e.g. if the file has moved), I’ll look into modifying libgambatte to allow the flags to be modified on reset.