Add an actual app icon
This first icon is a pixellated gradient, containing each of the colours that the Game Boy Color came in. Don’t mistake it for some colour mainpulation app…

(There’s also a minor and almost-authentic easter egg, for those of you who prefer the original GB 😉)

Anyway, I’m open to suggestions on how to improve it in the future.