Efficency improvements when in background
When in the background, Gambattye reduces its timer rate (as the controller doesn’t need to be polled), and when a window is occluded, it doesn’t bother rendering the image.

At present, there may be some minor audio skipping when transitioning into (or out of) the background, which could be improved in the future.

Also:
- Moves timer scheduling into rescheduleTimer(forStart:).
- Moves rendering into render().