Add basic icons for all file types
Add basic icons for GB, GBC, SAV, RTC, and even currently-unimplemented GQS files. These are in the typical extension-over-icon-on-paper style. This is starting to even feel like a proper Mac app(!)

I do have ideas on how to further improve the GB and GBC icons, but I think they’ll be for another time…