Link to GNU for licence, and update README
In the about box, now link to the GNU’s copy of the GPL, rather than to the LICENCE file in my own GitHub repo.

As this should stay up for the foreseeable future, I’ve also removed the address for the FSF. If this causes any issues, let me know.

Also, reference Gatekeeper in the README’s installation instructions. Just to handle any potential problems there. (As much as I’d like to properly sign Gambattye, I’m not ready to shell out money yet…)