Add Touch Bar support for loading/saving states
Let’s actually give the Touch Bar something to do! This lets you identify and load states nice and quickly.

It works in a similar fashion to the keyboard shortcuts, in that you tap a state to load it, and tap it while holding Option to save it.

In order to accomplish this sensibly, NSNotifications are now being used to load and save states, instead of closures.