Add support for original GB boot ROMs
If you happen to have the boot ROM for an original Game Boy lying around, Gambattye can show you the classic scroll and ‘ba-ding’ sound effect.

This can be set up from the new ‘Advanced’ section in the preferences.

To let the Advanced preferences have the Advanced icon, the Controls section now use the app icon as a stopgap.