Add support for GBC boot ROMs
If you have a copy of it lying around, this lets Gambattye show you the classic colourfulness and ‘ba-ding’ sound effect.

GBC boot ROMs will also be used when the GBA bit setting is set, which should probably be changed in the future.

I’ve also cleaned up some of the boot ROM-related code in general, including clearing up of notification observers.

Fixes #2.