Always use the system apperance
Ah, yes, the update that I’d been meaning to do since Mojave came out!

So, if you use dark mode, the app will be dark, and if not, the app will be light.

The only part that actually caused any problems is the KeyRecorderControl, due to its background colour always being white. I should really just replace it, seeing as it seems to be more trouble than it’s worth at the moment…