Update libgambatte
Integrates a variety of accuracy tweaks that have been made upstream, and reverts a few things that I didn’t think needed changing after all.

The only thing that directly affects the code of Gambattye is that we now have to flip the red and blue channels in the video buffer, instead of having libgambatte do this in the first place.